rename pandas assets to dynamic
