+++ great_expectations/datasource/fluent/pandas_dbfs_datasource.pyi
@@ -6,7 +6,7 @@ from typing import Hashable, Iterable, Literal, Optional, Sequence, Union
 from great_expectations._docs_decorators import public_api as public_api
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.datasource.fluent import PandasFilesystemDatasource, Sorter
-from great_expectations.datasource.fluent.data_asset.path.pandas.pandas import (
+from great_expectations.datasource.fluent.data_asset.path.pandas.dynamic_assets import (
     CSVAsset,
     ExcelAsset,
     FeatherAsset,
