+++ tests/datasource/fluent/test_pandas_azure_blob_storage_datasource.py
@@ -16,7 +16,7 @@ from great_expectations.datasource.fluent.config_str import ConfigStr
 from great_expectations.datasource.fluent.data_asset.path.file_path_data_asset import (
     _FilePathDataAsset,
 )
-from great_expectations.datasource.fluent.data_asset.path.pandas.pandas import CSVAsset
+from great_expectations.datasource.fluent.data_asset.path.pandas.dynamic_assets import CSVAsset
 from great_expectations.datasource.fluent.data_connector import (
     AzureBlobStorageDataConnector,
 )
