+++ tests/datasource/fluent/test_pandas_filesystem_datasource.py
@@ -19,7 +19,10 @@ from great_expectations.datasource.fluent import PandasFilesystemDatasource
 from great_expectations.datasource.fluent.data_asset.path.file_path_data_asset import (
     _FilePathDataAsset,
 )
-from great_expectations.datasource.fluent.data_asset.path.pandas.pandas import CSVAsset, JSONAsset
+from great_expectations.datasource.fluent.data_asset.path.pandas.dynamic_assets import (
+    CSVAsset,
+    JSONAsset,
+)
 from great_expectations.datasource.fluent.data_connector import (
     FilesystemDataConnector,
 )
