+++ tests/datasource/fluent/test_pandas_google_cloud_storage_datasource.py
@@ -18,7 +18,7 @@ from great_expectations.datasource.fluent import (
 from great_expectations.datasource.fluent.data_asset.path.file_path_data_asset import (
     _FilePathDataAsset,
 )
-from great_expectations.datasource.fluent.data_asset.path.pandas.pandas import CSVAsset
+from great_expectations.datasource.fluent.data_asset.path.pandas.dynamic_assets import CSVAsset
 from great_expectations.datasource.fluent.data_connector import (
     GoogleCloudStorageDataConnector,
 )
